+++ src/main/java/twilightforest/entity/ai/goal/RedcapShyGoal.java
@@ -89,7 +89,7 @@ public class RedcapShyGoal extends RedcapBaseGoal {
 		float rangle = (float) (Math.atan2(vecz, vecx));
 
 		// add a little, so he circles
-		rangle += rotation;
+		rangle += (float) rotation;
 
 		// figure out where we're headed from the target angle
 		double dx = Mth.cos(rangle) * radius;
